+++ src/commands/index.ts
@@ -1,7 +1,10 @@
 import { commands, ExtensionContext, window } from 'vscode';
+import { fix } from './_fixJSON';
 import { uglifyJson } from './_uglifyJson';
 import { beautifyJson } from './_beautifyJson';
-import { fix } from './_fixJSON';
+import { validateJson } from './_validateJson';
+import { escapeJson } from './_escapeJson';
+import { unescapeJson } from './_unescapeJson';
 
 module.exports = function (context: ExtensionContext) {
   const decoration = window.createTextEditorDecorationType({
@@ -14,4 +17,7 @@ module.exports = function (context: ExtensionContext) {
 
   context.subscriptions.push(commands.registerCommand("vscode-json-enhanced.uglifyJson", uglifyJson));
   context.subscriptions.push(commands.registerCommand("vscode-json-enhanced.beautifyJson", beautifyJson));
+  context.subscriptions.push(commands.registerCommand("vscode-json-enhanced.validateJson", validateJson));
+  context.subscriptions.push(commands.registerCommand("vscode-json-enhanced.escapeJson", escapeJson));
+  context.subscriptions.push(commands.registerCommand("vscode-json-enhanced.unescapeJson", unescapeJson));
 };