+++ CHANGELOG.md
@@ -1,11 +1,33 @@
 # Change Log
 
+## 1.0.5
+
+- Add `Escape JSON` command
+- Add `Unescape JSON` command
+- Add `Validate JSON` command
+
+> Escape JSON
+
+![](./doc/images/005.gif)
+
+> Unescape JSON
+
+![](./doc/images/006.gif)
+
+> Validate JSON
+
+![](./doc/images/007.gif)
+
 ## 1.0.4
 
 - Add `Uglify JSON` command
 - Add `Beautify JSON` command
 
+> Uglify JSON
+
 ![](./doc/images/003.gif)
+
+> Beautify JSON
 
 ![](./doc/images/004.gif)